command parser

+++ Robot/Ecran.h
@@ -6,11 +6,11 @@
 class Ecran
 {
     public:
-        void setup(byte _address);
+        void setup(int _address);
 
         bool send(String command);
         String receive(int quantity);
 
     private:
-        byte address;
+        int address;
 };
